+++ src/Shgif/Type.hs
@@ -15,6 +15,8 @@ module Shgif.Type (
     Format(..)
     , Shgif(..)
     , TimeStamp
+    , Updater
+    , Updatable
 
     -- * Small functions for Shgif
     , shgifToCanvas