+++ src/Shgif/Updater.hs
@@ -11,95 +11,82 @@ This module provides lots kind of __updater__.
 
 Each Updater has different update method.
 -}
+{-# LANGUAGE Rank2Types #-}
 module Shgif.Updater (
     -- * Normal
-      updateShgif
+      updateNormal
 
     -- With loop/reversed setting
-    , updateShgifNoLoop
-    , updateShgifReversed
-    , updateShgifReversedNoLoop
+    , updateNoLoop
+    , updateReversed
+    , updateReversedNoLoop
 
     -- * Flexible changing
-    , updateShgifTo
+    , updateTo
+    , setTickTo
 ) where
 import Shgif.Type.Internal
-import Control.Lens (over, set, (+~), (&), (^.))
+import Control.Lens (over, set, (+~), (&), (^.), (.~), view)
 
-
--- | Update 'Shgif''s internal tick state, which will affect frame rendering.
---
--- As 'updateShgif' has type `Shgif -> IO Shgif`, it can be called inside 'Brick.EventM' monad.
+-- | Update internal tick state, which will affect frame rendering.
 --
 -- This function __won't loop__ gif.
 --
 -- Use this if you want to show animation only once.
-updateShgifNoLoop :: Shgif -> IO Shgif
-updateShgifNoLoop shgif@(Shgif t a f w h tick ds c) = do
-    let updateTick | tick <= lastTimeStamp = over currentTick (+ 1)
-                   | otherwise             = id
-
-    newC <- shgifToCanvas $ updateTick shgif
-
-    return $ set canvas (Just newC) $ updateTick shgif
+updateNoLoop :: Updater
+updateNoLoop updatable = update updateTick updatable
     where
-        lastTimeStamp = maximum $ map fst ds
+        lastTimeStamp = getLastTimeStamp updatable
+        updateTick | (updatable^.getTick) <= lastTimeStamp = over getTick (+ 1)
+                   | otherwise                             = id
 
 
--- | Update 'Shgif''s internal tick state, which will affect frame rendering.  
---
--- As 'updateShgif' has type `Shgif -> IO Shgif`, it can be called inside brick's 'Brick.EventM' monad.
+-- | Update internal tick state, which will affect frame rendering.  
 --
 -- This function reverse and __won't loop__ gif.
 --
 -- Use this if you want to show reversed animation for only once.
-updateShgifReversedNoLoop :: Shgif -> IO Shgif
-updateShgifReversedNoLoop shgif = do
-    let updateTick | 0 < (shgif^.currentTick) = over currentTick (subtract 1)
+updateReversedNoLoop :: Updater
+updateReversedNoLoop updatable = update updateTick updatable
+    where
+        updateTick | 0 < (updatable^.getTick) = over getTick (subtract 1)
                    | otherwise                = id
-    newC <- shgifToCanvas $ updateTick shgif
-    return $ set canvas (Just newC) $ updateTick shgif
 
 
--- | Update 'Shgif''s internal tick state, which will affect frame rendering.  
---
--- As 'updateShgif' has type `Shgif -> IO Shgif`, it can be called inside brick's 'Brick.EventM' monad.
+-- | Update internal tick state, which will affect frame rendering.  
 --
 -- This function reverse gif.
 --
 -- Use this if you want to show reversed animation.
-updateShgifReversed :: Shgif -> IO Shgif
-updateShgifReversed shgif = do
-    let updateTick = set currentTick (repeat lastTimeStamp $ (shgif^.currentTick) - 1)
-    newC <- shgifToCanvas $ updateTick shgif
-    return $ set canvas (Just newC) $ updateTick shgif
+updateReversed :: Updater
+updateReversed updatable = update updateTick updatable
     where
-        lastTimeStamp = maximum $ map fst (shgif^.shgifData)
+        lastTimeStamp = getLastTimeStamp updatable
         -- https://docs.unity3d.com/ja/2019.2/ScriptReference/Mathf.Repeat.html
         repeat max val | val < 0   = max
-                       
--- | Update 'Shgif''s internal tick state, which will affect frame rendering.  
---
--- As 'updateShgif' has type `Shgif -> IO Shgif`, it can be called inside brick's 'Brick.EventM' monad.
---
-updateShgif :: Shgif -> IO Shgif
-updateShgif shgif = do
-    let updateTick = set currentTick (repeat lastTimeStamp $ (shgif^.currentTick) + 1)
-    newC <- shgifToCanvas $ updateTick shgif
-    return $ set canvas (Just newC) $ updateTick shgif
+        updateTick = set getTick (repeat lastTimeStamp $ (updatable^.getTick) - 1)
+
+-- | Update internal tick state, which will affect frame rendering.  
+updateNormal :: Updater
+updateNormal updatable = update updateTick updatable
     where
-        lastTimeStamp = maximum $ map fst (shgif^.shgifData)
+        lastTimeStamp = getLastTimeStamp updatable
         -- https://docs.unity3d.com/ja/2019.2/ScriptReference/Mathf.Repeat.html
         repeat max val | max < val = 0
                        | otherwise = val
+        updateTick = set getTick (repeat lastTimeStamp $ (updatable^.getTick) + 1)
 
 
--- | Update 'Shgif''s internal tick state to make it closer to given tick
-updateShgifTo :: Int -> Shgif -> IO Shgif
-updateShgifTo tick shgif  = do
-    let tickToAdd = case (shgif^.currentTick) `compare` tick of
+-- | Update internal tick state to make it closer to given tick
+updateTo :: Int -> Updater
+updateTo tick updatable  = update (getTick+~tickToAdd) updatable
+    where
+        tickToAdd = case (updatable^.getTick) `compare` tick of
                         LT -> 1
                         EQ -> 0
                         GT -> -1
-    newC <- shgifToCanvas (shgif&currentTick+~tickToAdd)
-    return $ set canvas (Just newC) $ over currentTick (+ tickToAdd) shgif
+
+
+-- | Set internal tick state to given tick.
+setTickTo :: Int -> Updater
+setTickTo tick = update (getTick.~tick)